% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lvl.R
\name{lvl}
\alias{lvl}
\title{Get code or level}
\usage{
lvl(a)
}
\arguments{
\item{a}{character variable of form "[code]. [level]"}
}
\description{
These functions extract the code (numeric) or the level (character) from a character variable of the form "[code]. [level]"
}
\examples{


lvl(c("1. one", "2. two"))
lvlval(c("1. one", "2. two"))
}
\keyword{BCIS}
