% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printing-statistics.R
\name{printinghelp}
\alias{printinghelp}
\alias{mround}
\title{Printing statistics nicely}
\usage{
mround(x, base)
}
\arguments{
\item{x}{Statistic}

\item{base}{rounds to the nearest 'base'}

\item{k}{Number of decminal places to round to}

\item{pval}{p-value (or any probability)}

\item{xLL, xUL}{lower and upper confidence limits}
}
\description{
Printing statistics nicely
}
\examples{

}
