% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_statistics.R
\name{population_estimates}
\alias{population_estimates}
\alias{var.p}
\alias{sd.p}
\title{Population estimates for var and sd}
\usage{
var.p(x, na.rm = FALSE)

sd.p(x, na.rm = FALSE)
}
\arguments{
\item{x}{a numeric vector.}

\item{na.rm}{logical. Should missing values be removed?}

\item{y}{NULL (default) or a vector.}
}
\description{
Population estimates for var and sd
}
\examples{

}
