% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printing-statistics.R
\name{printing}
\alias{printing}
\alias{mround}
\alias{specify_decimal}
\alias{print_pval}
\alias{print_estCIs}
\alias{print_CIs}
\title{Printing statistics nicely}
\usage{
mround(x, base)

specify_decimal(x, k)

print_pval(pval, k = 3)

print_estCIs(x, xLL, xUL, round = 1)

print_CIs(xLL, xUL, round = 1)
}
\arguments{
\item{x}{Statistic}

\item{base}{rounds to the nearest 'base'}

\item{k}{Number of decminal places to round to}

\item{pval}{p-value (or any probability)}

\item{xLL, xUL}{lower and upper confidence limits}
}
\description{
Printing statistics nicely
}
\examples{

}
