% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_tags.R
\name{summarise_tags}
\alias{summarise_tags}
\title{Summarise overlapping groups}
\usage{
summarise_tags(.tbl, .vars, ..., .removeF = FALSE)
}
\arguments{
\item{.tbl}{data}

\item{.vars}{unquoted variable names, passed to \link[dplyr]{vars}}

\item{...}{new variables to summarise, passed to \link[dplyr]{summarise}}

\item{.removeF}{defaults to `FALSE`. If `TRUE`, removes all values where output value is `FALSE`}
}
\description{
This extends dplyr::summarise to work with groups that overlap
}
\keyword{dplyr}
