% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lvl.R
\name{lvl}
\alias{lvl}
\alias{lvlval}
\title{Code-level functions}
\usage{
lvl(a)

lvlval(a)
}
\arguments{
\item{a}{character variable of form "[code]. [level]"}
}
\description{
These functions extract the code (numeric) or the level (character) from a character variable of the form "[code]. [level]"
}
