% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printing_statistics.R
\name{printing}
\alias{printing}
\alias{mround}
\alias{specify_decimal}
\alias{print_pval}
\alias{print_est1bracket}
\alias{print_est2bracket}
\alias{print_2bracket}
\alias{align_decimal}
\title{Printing statistics nicely}
\usage{
mround(x, base)

specify_decimal(x, k, trim = FALSE)

print_pval(pval, k = 3)

print_est1bracket(x, b, round = 1)

print_est2bracket(x, b1, b2, round = 1)

print_2bracket(b1, b2, round = 1)

align_decimal(x)
}
\arguments{
\item{x}{Statistic}

\item{base}{rounds to the nearest 'base'}

\item{k}{Number of decminal places to round to}

\item{pval}{p-value (or any probability)}

\item{b, b1, b2}{bracketed values}
}
\description{
Printing statistics nicely
}
\examples{

}
